[TASK] Round width and height calculations
Currently width and height calculations are floored.

The rounding difference can result in an image being loaded twice and minor layout shifts.

In my case the original image is 8420x4736 and the thumbnail gets resized to 80x44. Then the Image is first fetched in e.g. 1280x704 and then again with 1280x720  as the image space is slightly raised after the first load...

rounding instead of floor helps in most cases as a quick fix.