Only reload images in view on resize
This reduces the amount of images fetched during a resize
event and thus the overall bandwidth usage of the client
device.

Peaks in the server load due to image generation is
drastically reduced. Not all images viewed on the page
are requested at once, only the ones really in view.